move risk components to new file
