+++ src/vivarium_ciff_sam/components/__init__.py
@@ -2,8 +2,8 @@ from vivarium_ciff_sam.components.intervention import SQLNSIntervention, Wasting
 from vivarium_ciff_sam.components.mortality import Mortality
 from vivarium_ciff_sam.components.observers import (BirthObserver, CategoricalRiskObserver, DisabilityObserver,
                                                     DiseaseObserver, MortalityObserver)
-from vivarium_ciff_sam.components.lbwsg import LBWSGRisk, LBWSGRiskEffect, LBWSGSubRisk, LowBirthWeight, ShortGestation
-from vivarium_ciff_sam.components.maternal_malnutrition import AdditiveRiskEffect
+from vivarium_ciff_sam.components.lbwsg import LBWSGRiskEffect, LBWSGSubRisk, LowBirthWeight, ShortGestation
+from vivarium_ciff_sam.components.risk import AdditiveRiskEffect, RiskWithTracked
 from vivarium_ciff_sam.components.treatment import SQLNSTreatment, WastingTreatment
 from vivarium_ciff_sam.components.wasting import ChildWasting
 from vivarium_ciff_sam.components.x_factor import XFactorExposure, XFactorEffect
